+++ src/app/components/login/LoginComponent.ts
@@ -1,12 +1,11 @@
-import { Component, ElementRef, Injectable, Input } from "angular2/core";
-import { ROUTER_DIRECTIVES, ROUTER_PROVIDERS, Router, Instruction, RouterLink } from "angular2/router";
-import { CORE_DIRECTIVES, FORM_DIRECTIVES, FormBuilder, ControlGroup, AbstractControl, Control, NgIf, Validators, AbstractControl } from "angular2/common";
+import { Component, ElementRef, Input, Inject } from "angular2/core";
+import { ROUTER_DIRECTIVES, Router } from "angular2/router";
+import { CORE_DIRECTIVES, FORM_DIRECTIVES, FormBuilder, ControlGroup, AbstractControl, Validators } from "angular2/common";
 
-import Carbon from "carbonldp/Carbon";
-import * as Credentials from "carbonldp/Auth/Credentials";
+import { AuthService } from "angular2-carbonldp/services";
+
+import Credentials from "carbonldp/Auth/Credentials";
 import * as HTTP from "carbonldp/HTTP";
-import Cookies from "js-cookie";
-import AuthenticationToken from "carbonldp/Auth";
 
 import { ValidationService } from "app/components/validation-service/ValidationService";
 
@@ -18,18 +17,15 @@ import template from "./template.html!";
 @Component( {
 	selector: "login",
 	template: template,
-	directives: [ CORE_DIRECTIVES, ROUTER_DIRECTIVES, FORM_DIRECTIVES ]
+	directives: [ CORE_DIRECTIVES, ROUTER_DIRECTIVES, FORM_DIRECTIVES, ],
 } )
 export default class LoginComponent {
-	carbon:Carbon;
 	router:Router;
 	element:ElementRef;
-	private cookiesHandler:Cookies;
 
 	$element:JQuery;
 	$loginForm:JQuery;
 
-	submitting:boolean = false;
 	sending:boolean = false;
 	errorMessage:string = "";
 
@@ -37,32 +33,35 @@ export default class LoginComponent {
 	formBuilder:FormBuilder; // Validators
 	email:AbstractControl; // To make available the state of the input in the template
 	password:AbstractControl; // To make available the state of the input in the template
+	rememberMe:AbstractControl;
+	remember:boolean = true;
 
 	@Input() container:string|JQuery;
+	private authService:AuthService.Class;
 
-	constructor( public router:Router, element:ElementRef, formBuilder:FormBuilder, carbon:Carbon ) {
+	constructor( router:Router, element:ElementRef, formBuilder:FormBuilder, @Inject( AuthService.Token ) authService:AuthService.Class ) {
 		this.router = router;
 		this.element = element;
 		this.formBuilder = formBuilder;
-		this.carbon = carbon;
-		this.cookiesHandler = Cookies;
+		this.authService = authService;
 	}
 
 	ngOnInit():void {
 		this.$element = $( this.element.nativeElement );
 		this.$loginForm = this.$element.find( "form.loginForm" );
+		this.$loginForm.find( ".ui.checkbox" ).checkbox();
 		this.loginForm = this.formBuilder.group( {
 			email: [ "", Validators.compose( [ Validators.required, ValidationService.emailValidator ] ) ],
-			password: [ "", Validators.compose( [ Validators.required ] ) ]
+			password: [ "", Validators.compose( [ Validators.required ] ) ],
+			rememberMe: [ "", Validators.compose( [] ) ],
 		} );
 		this.email = this.loginForm.controls[ "email" ];
 		this.password = this.loginForm.controls[ "password" ];
+		this.rememberMe = this.loginForm.controls[ "rememberMe" ];
 	}
 
-	onSubmit( data:{ email:string, password:string }, $event:any ) {
+	onSubmit( data:{ email:string, password:string, rememberMe:boolean }, $event:any ):void {
 		$event.preventDefault();
-		this.submitting = true;
-		this.submitted = true;
 		this.sending = true;
 		this.errorMessage = "";
 		this.email.markAsTouched();
@@ -75,37 +74,52 @@ export default class LoginComponent {
 
 		let username:string = data.email;
 		let password:string = data.password;
+		let rememberMe:boolean = ! ! data.rememberMe;
 
-		this.carbon.auth.authenticate( username, password ).then(
-			( credential:Credentials ) => {
-				this.sending = false;
-				this.submitting = false;
-				let days:number = this.getDays( (new Date()), credential.expirationTime );
-				this.cookiesHandler.set( "carbon_jwt", credential, days );
-				this.router.navigate( [ "/AppDev" ] );
-			} ).catch( ( error:Error ) => {
-				this.sending = false;
-
-				switch ( true ) {
-					case error instanceof HTTP.Errors.UnauthorizedError:
-						this.errorMessage = "Wrong credentials";
-						break;
-					default:
-						this.errorMessage = "There was a problem processing the request";
-						break;
-				}
-			}
-		);
-		this.submitting = false;
-		this.errorMessage = "Service temporary unavailable.";
+		this.authService.login( username, password, rememberMe ).then( ( credential:Credentials ) => {
+			this.sending = false;
+			this.router.navigate( [ "/AppDev" ] );
+		} ).catch( ( error:HTTP.Errors.Error ) => {
+			this.sending = false;
+			this.setErrorMessage( error );
+		} );
 	}
 
 	getDays( firstDate:Date, lastDate:Date ):number {
 		// Discard the time and time-zone information
-		let utc1 = Date.UTC( firstDate.getFullYear(), firstDate.getMonth(), firstDate.getDate() );
-		let utc2 = Date.UTC( lastDate.getFullYear(), lastDate.getMonth(), lastDate.getDate() );
-		let ms_per_day = 1000 * 60 * 60 * 24;
-		return Math.floor( (utc2 - utc1) / ms_per_day );
+		let utc1:number = Date.UTC( firstDate.getFullYear(), firstDate.getMonth(), firstDate.getDate() );
+		let utc2:number = Date.UTC( lastDate.getFullYear(), lastDate.getMonth(), lastDate.getDate() );
+		let msPerDay:number = 1000 * 60 * 60 * 24;
+		return Math.floor( (utc2 - utc1) / msPerDay );
+	}
+
+	setErrorMessage( error:HTTP.Errors.Error ):void {
+		switch ( true ) {
+			case error instanceof HTTP.Errors.ForbiddenError:
+				this.errorMessage = "Denied Access.";
+				break;
+			case error instanceof HTTP.Errors.UnauthorizedError:
+				this.errorMessage = "Wrong credentials.";
+				break;
+			case error instanceof HTTP.Errors.BadGatewayError:
+				this.errorMessage = "An error occurred while trying to login. Please try again later. Error: " + error.response.status;
+				break;
+			case error instanceof HTTP.Errors.GatewayTimeoutError:
+				this.errorMessage = "An error occurred while trying to login. Please try again later. Error: " + error.response.status;
+				break;
+			case error instanceof HTTP.Errors.InternalServerErrorError:
+				this.errorMessage = "An error occurred while trying to login. Please try again later. Error: " + error.response.status;
+				break;
+			case error instanceof HTTP.Errors.UnknownError:
+				this.errorMessage = "An error occurred while trying to login. Please try again later. Error: " + error.response.status;
+				break;
+			case error instanceof HTTP.Errors.ServiceUnavailableError:
+				this.errorMessage = "Service currently unavailable.";
+				break;
+			default:
+				this.errorMessage = "There was a problem processing the request. Error: " + error.response.status;
+				break;
+		}
 	}
 
 	shakeForm():void {
@@ -115,7 +129,7 @@ export default class LoginComponent {
 		}
 		if ( target ) {
 			target.transition( {
-				animation: "shake"
+				animation: "shake",
 			} );
 		}
 	}